remove sever linux  IPSender
